SCR/SCRAPING/requirements.txt
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
dask==2024.1.0
fsspec==2023.12.2
google-api-core==2.15.0
google-auth==2.26.2
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
idna==3.6
importlib-metadata==7.0.1
locket==1.0.0
numpy==1.26.3
packaging==23.2
pandas==2.1.4
partd==1.4.1
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
setuptools==68.2.2
six==1.16.0
toolz==0.12.0
tzdata==2023.4
urllib3==2.1.0
wheel==0.41.2
zipp==3.17.0
